+++ src/graphql/client/derives.rs
@@ -616,11 +616,3 @@ pub struct Export;
     response_derives = "Clone, Default, PartialEq"
 )]
 pub struct Statistics;
-
-#[derive(GraphQLQuery)]
-#[graphql(
-    schema_path = "src/graphql/client/schema/schema.graphql",
-    query_path = "src/graphql/client/schema/tsv_formatted_raw_events.graphql",
-    response_derives = "Clone, Default, PartialEq"
-)]
-pub struct TsvFormattedRawEvents;